requirements.txt
aiohttp
pydantic~=2.10.3
psutil>=5.8.0
openai
anthropic
qq-botpy
chardet~=5.1.0
Pillow
beautifulsoup4
googlesearch-python
readability-lxml
quart
lxml_html_clean
colorlog
aiocqhttp
pyjwt
apscheduler
docstring_parser
aiodocker
silk-python
lark-oapi
ormsgpack
cryptography
dashscope
python-telegram-bot
wechatpy
dingtalk-stream
defusedxml
mcp
certifi
pip
telegramify-markdown
google-genai
click
filelock
watchfiles
websockets
faiss-cpu
aiosqlite
nh3
py-cord>=2.6.1
slack-sdk
pydub
opencv-python